+++ src/common/index.jsx
@@ -1,6 +1,6 @@
-export { default as Button } from './Button/Button'
-export { default as ArticleSection } from './ArticleSection/ArticleSection'
-export { default as Navbar } from './Navbar/Navbar'
-export { default as Form } from './Form/Form'
-export { default as PersonCard } from './PersonCard/PersonCard'
-export { default as ScrollDown } from './ScrollDown/ScrollDown'
+export { default as Button } from './Button'
+export { default as ArticleSection } from './ArticleSection'
+export { default as Navbar } from './Navbar'
+export { default as Form } from './Form'
+export { default as PersonCard } from './PersonCard'
+export { default as ScrollDown } from './ScrollDown'
